Address review: isolate conflict-time skipping in a self-typed trait
Move buildDataSkippingPredicate / filterFilesByDataSkipping /
filterFilesMatchingAnyReadPredicate out of DataSkippingReaderBase into a new
self-typed trait ConflictDataSkippingReader (self: DataSkippingReaderBase),
mixed into the base. Existing data-skipping code is left untouched:
withStatsInternal0 is reverted to its original inline form, so the only change
to DataSkippingReader.scala is the one-line trait mixin.

parseAndDecodeStats is duplicated (trait-private) rather than shared with
withStatsInternal0, keeping the feature a fully isolated additive unit per
review feedback.
