+++ spark/src/main/scala/org/apache/spark/sql/delta/stats/DataSkippingReader.scala
@@ -20,7 +20,6 @@ package org.apache.spark.sql.delta.stats
 import java.io.Closeable
 
 import scala.collection.mutable.ArrayBuffer
-import scala.util.control.NonFatal
 
 import org.apache.spark.sql.delta.skipping.clustering.{ClusteredTableUtils, ClusteringColumnInfo}
 import org.apache.spark.sql.delta.ClassicColumnConversions._
@@ -265,6 +264,7 @@ trait DataSkippingReaderBase
   with StatisticsCollection
   with ReadsMetadataFields
   with StateCache
+  with ConflictDataSkippingReader
   with DeltaLogging {
 
   import DataSkippingReader._
@@ -287,28 +287,21 @@ trait DataSkippingReaderBase
         DeltaSQLConf.DELTA_DATASKIPPING_PARTITION_LIKE_FILTERS_ADDITIONAL_SUPPORTED_EXPRESSIONS)
       .toSet.flatMap((exprs: String) => exprs.split(","))
 
-  /**
-   * Parses (and, when the schema contains VariantType, Z85-decodes) the JSON `statsCol` into the
-   * `statsSchema` struct. Shared by [[withStatsInternal0]] and conflict-detection data skipping
-   * ([[filterFilesByDataSkipping]]) so both parse stats the same way.
-   */
-  private def parseAndDecodeStats(statsCol: Column): Column = {
-    val parsedStats = from_json(statsCol, statsSchema)
+  /** Returns a DataFrame expression to obtain a list of files with parsed statistics. */
+  private def withStatsInternal0: DataFrame = {
+    val parsedStats = from_json(col("stats"), statsSchema)
     // Only use DecodeNestedZ85EncodedVariant if the schema contains VariantType.
     // This avoids performance overhead for tables without variant columns.
     // `DecodeNestedZ85EncodedVariant` is a temporary workaround since the Spark 4.1 from_json
     // expression has no way to decode a VariantVal from an encoded Z85 string.
     // TODO: Add Z85 decoding to Variant in Spark 4.2 and use that from_json option here.
-    if (SchemaUtils.checkForVariantTypeColumnsRecursively(statsSchema)) {
+    val decodedStats = if (SchemaUtils.checkForVariantTypeColumnsRecursively(statsSchema)) {
       Column(DecodeNestedZ85EncodedVariant(parsedStats.expr))
     } else {
       parsedStats
     }
+    allFiles.withColumn("stats", decodedStats)
   }
-
-  /** Returns a DataFrame expression to obtain a list of files with parsed statistics. */
-  private def withStatsInternal0: DataFrame =
-    allFiles.withColumn("stats", parseAndDecodeStats(col("stats")))
 
   private lazy val withStatsCache =
     cacheDS(withStatsInternal0, s"Delta Table State with Stats #$version - $redactedPath")
@@ -647,112 +640,6 @@ trait DataSkippingReaderBase
       convertDataFrameToAddFiles(df)
     }
     files.toSeq -> Seq(DataSize(totalSize), DataSize(partitionSize), DataSize(scanSize))
-  }
-
-  /**
-   * Builds a single [[DataSkippingPredicate]] (skipping expression + the stats it references) from
-   * `dataFilters`, mirroring [[filesForScan]]'s eligibility filtering, per-filter construction and
-   * conjunction fold. Returns None when stats skipping is unavailable or no eligible filter yields
-   * a predicate. The filters are AND-combined, so callers must pass filters from a single logical
-   * read (predicates from independent reads have OR, not AND, semantics).
-   */
-  private[delta] def buildDataSkippingPredicate(
-      dataFilters: Seq[Expression]): Option[DataSkippingPredicate] = {
-    import DeltaTableUtils._
-    if (!useStats) return None
-    // Mirror filesForScan eligibility: drop subquery / non-deterministic / metadata filters, so we
-    // never build a skipping predicate that could wrongly exclude a matching file.
-    val eligibleFilters = dataFilters.filterNot { f =>
-      containsSubquery(f) || !f.deterministic || f.exists {
-        case MetadataAttribute(_) => true
-        case _ => false
-      }
-    }
-    val constructDataFilters = new DataFiltersBuilder(
-      spark = spark,
-      dataSkippingType = DeltaDataSkippingType.dataSkippingOnlyV1,
-      getStatsColumnOpt = (s: StatsColumn) => getStatsColumnOpt(s))
-    eligibleFilters
-      .flatMap(f => constructDataFilters(f))
-      .reduceOption((skip1, skip2) => DataSkippingPredicate(
-        skip1.expr && skip2.expr, skip1.referencedStats ++ skip2.referencedStats))
-  }
-
-  /**
-   * Conflict-detection helper (reader-side data skipping): returns the subset of `files` whose
-   * statistics do NOT prove they fail `dataFilters` -- i.e. the files that could still match and
-   * therefore must be treated as conflicts. Files with missing/insufficient stats (or when skipping
-   * is unavailable) are kept.
-   *
-   * `dataFilters` must come from a single logical read: they are AND-combined via
-   * [[buildDataSkippingPredicate]]. This is the one-read case of
-   * [[filterFilesMatchingAnyReadPredicate]]; callers with multiple independent reads should use
-   * that method so all reads are evaluated in a single Spark job.
-   */
-  private[delta] def filterFilesByDataSkipping(
-      files: Seq[AddFile],
-      dataFilters: Seq[Expression]): Seq[AddFile] =
-    filterFilesMatchingAnyReadPredicate(files, Seq(dataFilters))
-
-  /**
-   * Conflict-detection helper (reader-side data skipping) over several INDEPENDENT reads: returns
-   * the subset of `files` that could still match ANY read and therefore must be treated as
-   * conflicts. Each inner `Seq[Expression]` is one logical read's data filters (AND-combined via
-   * [[buildDataSkippingPredicate]]); the reads are OR-combined, matching read semantics -- a file
-   * is a candidate if it could match any one of them.
-   *
-   * All reads are evaluated in a SINGLE Spark job: the per-read skipping predicates are OR-ed
-   * together into one `where` clause, rather than filtering per read and unioning the survivors
-   * (which launched one job per read). Note we cannot instead flatten every read's filters into one
-   * predicate -- that would AND them (read1 AND read2), the opposite of the OR we need.
-   *
-   * One-way safe: each read contributes `expr || !verifyStatsForFilter(...)` exactly as
-   * [[getDataSkippedFiles]], so a file is dropped only when its stats *prove* it fails EVERY read
-   * -- a real conflict is never a false negative. Files with missing/insufficient stats, a
-   * read with no usable skipping predicate (empty / ineligible filters -> matches everything), or a
-   * table without stats are all kept. Returns the original [[AddFile]]s (matched by path).
-   *
-   * Fail-safe: skipping here is a pure optimization over correct (conservative) conflict detection,
-   * so if building or evaluating the predicate throws we fall back to the default behavior of
-   * keeping all `files` as conflict candidates rather than failing the commit.
-   */
-  private[delta] def filterFilesMatchingAnyReadPredicate(
-      files: Seq[AddFile],
-      dataFiltersPerRead: Seq[Seq[Expression]]): Seq[AddFile] = {
-    import org.apache.spark.sql.delta.implicits._
-    if (files.isEmpty || dataFiltersPerRead.isEmpty || schema.isEmpty) return files
-    try {
-      // One skipping predicate per read. A read with no usable predicate (empty or ineligible
-      // filters) matches every file -> nothing can be skipped, so keep all files without a job.
-      val perReadPredicates = dataFiltersPerRead.map { dataFilters =>
-        if (dataFilters.isEmpty) None else buildDataSkippingPredicate(dataFilters)
-      }
-      if (perReadPredicates.exists(_.isEmpty)) return files
-      // Survive if the file could match ANY read. Per read, `expr || !verifyStatsForFilter(...)`
-      // keeps any file whose referenced stats are missing/NULL (mirrors getDataSkippedFiles): only
-      // skip when stats prove no match. OR the reads so a match against any one keeps the file.
-      val survivorCondition = perReadPredicates.flatten
-        .map(pred => pred.expr || !verifyStatsForFilter(pred.referencedStats))
-        .reduce(_ || _)
-      val survivingPaths = recordFrameProfile(
-          "Delta", "DataSkippingReader.filterFilesMatchingAnyReadPredicate") {
-        files.toDF(spark)
-          .withColumn("stats", parseAndDecodeStats(col("stats")))
-          .where(survivorCondition)
-          .select("path")
-          .collect()
-          .map(_.getString(0))
-          .toSet
-      }
-      files.filter(f => survivingPaths.contains(f.path))
-    } catch {
-      case NonFatal(e) =>
-        // Optimization only: never let a skipping failure abort a commit. Fall back to the default
-        // (feature-off) behavior of treating every added file as a conflict candidate.
-        logWarning(log"Conflict-time data skipping failed to evaluate; falling back to treating " +
-          log"all added files as conflict candidates", e)
-        files
-    }
   }
 
   private def getCorrectDataSkippingType(
